This restores the branch at the state of r47415.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/fix_floatingpoint_calculator@47425 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cfa/simplification/ExpressionSimplificationVisitor.java
@@ -286,7 +286,7 @@ public class ExpressionSimplificationVisitor
             return new CFloatLiteralExpression(
                 loc, machineModel, exprType, negatedValue.floatingPointValue());
           default:
-            // return the original expression below
+            // fall-through and return the original expression
         }
 
       } else if (unaryOperator == UnaryOperator.TILDE
@@ -332,8 +332,8 @@ public class ExpressionSimplificationVisitor
 
     // enum constant
     if (decl instanceof CEnumerator) {
-      final BigInteger v = ((CEnumerator) decl).getValue();
-      return new CIntegerLiteralExpression(expr.getFileLocation(), type, v);
+      final long v = ((CEnumerator) decl).getValue();
+      return new CIntegerLiteralExpression(expr.getFileLocation(), type, BigInteger.valueOf(v));
     }
 
     // const variable, inline initializer
@@ -361,7 +361,7 @@ public class ExpressionSimplificationVisitor
               return new CFloatLiteralExpression(
                   expr.getFileLocation(), machineModel, type, castValue);
             default:
-              // return the original expression below
+              // fall-through and return the original expression
           }
         }
       }
